Use a single CustomPainter and enable it to receive tap-gestures.
This was previously trying to use two painters, one for painting the
puzzle background and and the other to receive tap-gestures and display
the corresponding user-moves. These two are now combined and the new
PuzzlePainter class is re-defined so as to allow a repaint whenever
there is a tap-gesture. Also, the paint() procedure now re-draws every
detail of the puzzle-state every time it repaints, as per the Flutter
convention for painting a Canvas object.