Rejig model/view hierarchy and interfaces. Add Killer & Mathdoku types.
User messages have to be issued in the Puzzle View, with a Widget as
parent. Also, it is necessary to have an executeAfterBuild() event to
take care of any Model changes that are detected during widget building
and painting, such as the user finishing the puzzle. An InheritedWidget
was added to try and make access to Model data smoother in the View.
Finally, an initial version of Mathdoku and Killer types was added, but
without the DLXSolver to detect puzzles with multiple solutions.