Avoid changing Puzzle model from Canvas. Start introducing Provider.
When the PointerDown position was passed into the CustomPaint Canvas and
then used to update the Puzzle model, one PointerDown event could be
processed many times if Flutter itself triggered repaints, such as when
hovering near an Icon Button. This made it difficult to "undo" a tap by
simply tapping again. The Puzzle is now updated elsewhere and the Canvas
just reads and repaints the Puzzle state (in its entirety). This also
paves the way for bringing in Provider, to manage updating the widgets
whenever the model is changed, by whichever of several methods.