Fix the startup for Puzzle and PaintingSpecs. Adjust some View details.
The Puzzle model, the PaintingSpecs and the PuzzleMap are instantiated
together, at a level just above PuzzleView in the widget tree, leaving
PuzzleView the small task of recording the Orientation of the device or
window in PaintingSpecs. This allows the model(s) to be unaffected even
if the entire PuzzleView sub-tree is re-painted (e.g. during a window
resize or a device re-orientation). Ordinarily, just the PuzzlePainter
and Canvas are re-painted when the Puzzle model changes during play.